Revert "The payment request is not done for the invoice balance any more, but for the invoice total."
This reverts commit 379983ca665677d307abbb4fd13f4e615f0002b0.

This was a bad idea because: if there are many invoices with balances, with the latest having a total for the sum of the balances, it is true that sending a request for a payment of the balance of only this last invoice will not pay all the previous ones, but the previous ones do have the chance of getting paid by retries. So instead of having one payment done for the last invoice that pays them all, there will be many, one per invoice. This is a resonable outcome, and avoid many potential problems of asking for a payment that is bigger of what the customer owes.